treewide: merge upstream v4.3.0 release
Merge from zephyr v4.3 release commit
3568e1b6d5cdd51a6b964a2a1d6d29200fea2056.

+++ drivers/sensor/ams/tsl2591/tsl2591.c
@@ -192,7 +192,7 @@ static int tsl2591_set_threshold(const struct device *dev, enum sensor_attribute
 	int ret;
 
 	/* Convert from relative strength of visible light to raw value */
-	cpl = data->atime * data->again;
+	cpl = (uint32_t)data->atime * data->again;
 	raw = ((val->val1 * cpl) / TSL2591_LUX_DF) +
 	      ((val->val2 * cpl) / (1000000U * TSL2591_LUX_DF));
